+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 **Enhancements:**
 
+- Changed Event ID Statistics option to Event ID Metrics option. (`-s, --statistics`  -> `-M, --metrics`) (#706) (@hitenkoku)
+  (Note: `statistics_event_info.txt` was changed to `event_id_info.txt`.)
 - Added display new version of Hayabusa when updating. (#710) (@hitenkoku)
 
 **Bug Fixes:**